requirements.txt
Flask==2.0.3
Flask-Session==0.4.0
Werkzeug==2.2.2
mysql-connector-python==8.0.28